requirements.txt
flask
flask-wtf
flask-sqlalchemy
mysql-connector
mysql-connector-python
mysql-connector-python-rf